refactor: move station classes
