--- src/main/java/nextstep/subway/Station.java
@@ -1,54 +1,0 @@
-package nextstep.subway;
-
-import java.util.Objects;
-
-import javax.persistence.*;
-
-@Entity
-public class Station {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(length = 20, nullable = false)
-    private String name;
-
-    public Station() {
-    }
-
-    public Station(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Station(String name) {
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public boolean equals(Object object) {
-        if (this == object) {
-            return true;
-        }
-
-        if (object == null || getClass() != object.getClass()) {
-            return false;
-        }
-
-        Station station = (Station)object;
-        return Objects.equals(name, station.name);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name);
-    }
-}
