--- src/main/java/nextstep/subway/StationRepository.java
@@ -1,6 +1,0 @@
-package nextstep.subway;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StationRepository extends JpaRepository<Station, Long> {
-}
