--- src/main/java/nextstep/subway/StationRequest.java
@@ -1,9 +1,0 @@
-package nextstep.subway;
-
-public class StationRequest {
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-}
