+++ src/main/java/nextstep/subway/application/DefaultStationService.java
@@ -1,0 +1,47 @@
+package nextstep.subway.application;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import nextstep.subway.application.dto.StationRequest;
+import nextstep.subway.application.dto.StationResponse;
+import nextstep.subway.domain.model.Station;
+import nextstep.subway.domain.repository.StationRepository;
+import nextstep.subway.domain.service.StationService;
+
+@Service
+@Transactional(readOnly = true)
+public class DefaultStationService implements StationService {
+    private StationRepository stationRepository;
+
+    public DefaultStationService(StationRepository stationRepository) {
+        this.stationRepository = stationRepository;
+    }
+
+    @Transactional
+    public StationResponse saveStation(StationRequest stationRequest) {
+        Station station = stationRepository.save(new Station(stationRequest.getName()));
+        return createStationResponse(station);
+    }
+
+    public List<StationResponse> findAllStations() {
+        return stationRepository.findAll().stream()
+                .map(this::createStationResponse)
+                .collect(Collectors.toList());
+    }
+
+    @Transactional
+    public void deleteStationById(Long id) {
+        stationRepository.deleteById(id);
+    }
+
+    private StationResponse createStationResponse(Station station) {
+        return new StationResponse(
+                station.getId(),
+                station.getName()
+        );
+    }
+}
