+++ src/main/java/nextstep/subway/application/dto/StationRequest.java
@@ -1,0 +1,9 @@
+package nextstep.subway.application.dto;
+
+public class StationRequest {
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+}
