+++ src/main/java/nextstep/subway/application/dto/StationResponse.java
@@ -1,0 +1,19 @@
+package nextstep.subway.application.dto;
+
+public class StationResponse {
+    private Long id;
+    private String name;
+
+    public StationResponse(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
