+++ src/main/java/nextstep/subway/domain/service/LineQueryService.java
@@ -1,0 +1,11 @@
+package nextstep.subway.domain.service;
+
+import java.util.List;
+
+import nextstep.subway.application.dto.LineResponse;
+
+public interface LineQueryService {
+    LineResponse findLineById(Long id);
+
+    List<LineResponse> findAllLines();
+}
