+++ src/main/java/nextstep/subway/domain/service/StationService.java
@@ -1,0 +1,14 @@
+package nextstep.subway.domain.service;
+
+import java.util.List;
+
+import nextstep.subway.application.dto.StationRequest;
+import nextstep.subway.application.dto.StationResponse;
+
+public interface StationService {
+    StationResponse saveStation(StationRequest stationRequest);
+
+    List<StationResponse> findAllStations();
+
+    void deleteStationById(Long id);
+}
