--- src/main/java/nextstep/subway/line/domain/LineRepository.java
@@ -1,16 +1,0 @@
-package nextstep.subway.line.domain;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface LineRepository {
-    Optional<Line> findById(Long id);
-
-    List<Line> findAll();
-
-    Line save(Line line);
-
-    void deleteById(Long id);
-
-    boolean existsById(Long id);
-}
