+++ src/test/java/nextstep/subway/acceptance/TestFixture.java
@@ -8,8 +8,8 @@ import org.springframework.http.MediaType;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.subway.line.application.dto.LineRequest;
-import nextstep.subway.line.application.dto.SectionRequest;
+import nextstep.subway.application.dto.LineRequest;
+import nextstep.subway.application.dto.SectionRequest;
 
 public class TestFixture {
